SFC parseComponent pads complete content with spaces (#5059)
* SFC parseComponent pads content with spaces

when `{ pad: true }` is provided. That is, all content is converted to
spaces. Previously, each line was truncated to "//". The new padding
method works better with character-oriented tools that calculate
positions by distance from the beginning of the file instead of by line
number.

* Made parseComponent's pad support "line" | "space"

Also still supports true for backward compatibility. True is the same as
"line".